requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
aiofiles==23.2.0